+++ Engine/AudioHelper.hpp
@@ -9,17 +9,17 @@ private:
     AudioHelper() = default;
 
 public:
-    static float BGMVolume;
-    static float SFXVolume;
+    static double BGMVolume;
+    static double SFXVolume;
     static ALLEGRO_SAMPLE_ID PlayAudio(const std::string &audio);
-    static ALLEGRO_SAMPLE_ID PlayAudio(const std::string &audio, const float& volumeScale);
+    static ALLEGRO_SAMPLE_ID PlayAudio(const std::string &audio, const double& volumeScale);
     static ALLEGRO_SAMPLE_ID PlayBGM(const std::string &audio);
     static void StopBGM(ALLEGRO_SAMPLE_ID sample);
-    static std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> PlaySample(const std::string &audio, bool loop = false, float volume = 1, float position = 0);
+    static std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> PlaySample(const std::string &audio, bool loop = false, double volume = 1, double position = 0);
     static void StopSample(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance);
-    static void ChangeSampleVolume(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, float volume);
-    static void ChangeSamplePosition(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, float position);
-    static void ChangeSampleSpeed(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, float speed);
+    static void ChangeSampleVolume(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, double volume);
+    static void ChangeSamplePosition(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, double position);
+    static void ChangeSampleSpeed(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, double speed);
     static unsigned int GetSampleLength(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance);
     static double GetSamplePosition(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance);
 };
