+++ Objects/Hold.h
@@ -5,7 +5,7 @@
 class Hold : public HitObject
 {
 public:
-    Hold(int x, int time, int type, int endTime, float positionY, float positionEndY, const std::string& hitSoundPath) :
+    Hold(int x, int time, int type, int endTime, double positionY, double positionEndY, const std::string& hitSoundPath) :
         HitObject(x, time, type, endTime, positionY, positionEndY, hitSoundPath) { ; }
     void OnKeyDown() override;
     void OnKeyUp() override;
